backend/requirements.txt
# flask==2.3.3
# flask-cors==4.0.0
# flask-caching==2.1.0
# pyvi==0.1.1
# sentence-transformers==2.2.2
# faiss-cpu==1.7.4
# numpy==1.26.0
# requests==2.31.0
# markdown==3.5.1

flask
flask-cors
flask-caching
pyvi
sentence-transformers
faiss-cpu
numpy
requests
markdown